Refactor test_monitor.py: Move db handler tests into own test source file test_monitor_handler.py
